backend/requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.7
Django==6.0.4
django-cors-headers==4.9.0
django-picklefield==3.4.0
django-q2==1.9.0
djangorestframework==3.17.1
djangorestframework-camel-case==1.4.2
greenlet==3.4.0
playwright==1.58.0
psycopg2-binary==2.9.11
pycparser==3.0
pyee==13.0.1
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
